pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py-atc-ble-oepl"
version = "0.1.0"
description = "Python module for interacting with ATC BLE firmware on OEPL e-paper tags"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "g4bri3lDev", email = "admin@g4bri3l.de"}
]

keywords = []  # TODO: Add relevant keywords
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "bleak>=0.21.0",
    "bleak-retry-connector>=3.5.0",
    "epaper-dithering>=0.1.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=0.23.0",
]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10"
]

[project.urls]
Homepage = "https://opendisplay.org"
Repository = "https://github.com/OpenDisplay-org/py-atc-ble-oepl"
Documentation = "https://github.com/OpenDisplay-org/py-atc-ble-oepl#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/py_atc_ble_oepl"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
